# storage/engine: minor pebble timeseries benchmark fix