Use an in-memory instance to match RocksDB's equivalent benchmark, and to
avoid writing more merge operands per key than intended.

Release note: None