@ajkr 2019-10-28
This fixes the single merge operand performance but not the multi-operand performance.

```
name                                                    old time/op  new time/op   delta
MVCCGetMergedTimeSeries/numKeys=1/mergesPerKey=1-6      6.42µs ± 3%   1.99µs ± 2%   -68.95%  (p=0.000 n=10+9)
MVCCGetMergedTimeSeries/numKeys=1/mergesPerKey=16-6     63.9µs ± 2%   95.2µs ± 1%   +49.15%  (p=0.000 n=9+10)
MVCCGetMergedTimeSeries/numKeys=1/mergesPerKey=256-6    9.80ms ± 3%  12.98ms ± 1%   +32.47%  (p=0.000 n=10+10)
MVCCGetMergedTimeSeries/numKeys=16/mergesPerKey=1-6     6.71µs ± 2%   2.50µs ± 8%   -62.80%  (p=0.000 n=9+10)
MVCCGetMergedTimeSeries/numKeys=16/mergesPerKey=16-6    65.7µs ± 3%  184.5µs ± 3%  +180.71%  (p=0.000 n=10+10)
MVCCGetMergedTimeSeries/numKeys=16/mergesPerKey=256-6   9.85ms ± 2%  25.05ms ± 5%  +154.16%  (p=0.000 n=10+10)
MVCCGetMergedTimeSeries/numKeys=256/mergesPerKey=1-6    7.15µs ± 2%   2.95µs ± 3%   -58.76%  (p=0.000 n=10+10)
MVCCGetMergedTimeSeries/numKeys=256/mergesPerKey=16-6   65.0µs ± 2%  189.4µs ± 1%  +191.64%  (p=0.000 n=10+9)
MVCCGetMergedTimeSeries/numKeys=256/mergesPerKey=256-6  9.90ms ± 3%  25.97ms ± 1%  +162.25%  (p=0.000 n=10+8)
```